Added server node.js implementation
server.js opens a socket to the MQTT broker on port 1883 and listens for web connections on port 8083. Upon connection, it creates a new client and forwards all data between the client and the broker over Socket.IO (the main benefit being interoperability).
server.sh is an runner script that starts the node server on boot.